src/main/java/com/twu/biblioteca/logic/Book.java
package com.twu.biblioteca.logic;

import java.util.Objects;

public class Book {
    private final String name;
    private final String author;
    private final int publicationYear;

    public Book(String name, String author, int publicationYear) {
        this.name = name;
        this.author = author;
        this.publicationYear = publicationYear;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Book book = (Book) o;
        return publicationYear == book.publicationYear &&
                Objects.equals(name, book.name) &&
                Objects.equals(author, book.author);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name, author, publicationYear);
    }

    public String getName() {
        return name;
    }

    public String display(String format) {
        return String.format(format, name, author, publicationYear);
    }

}
